refactore: update ticks counter system

+++ inc/managers/save_manager.hpp
@@ -122,7 +122,7 @@ class SaveManager {
 
       TYRA_LOG("Loading tick state...");
       // These are global scope variables;
-      g_ticksCounter = savedData["tickState"]["g_ticksCounter"].get<float>();
+      g_ticksCounter = savedData["tickState"]["g_ticksCounter"].get<uint32_t>();
       elapsedRealTime = savedData["tickState"]["elapsedRealTime"].get<double>();
       ticksDayCounter =
           savedData["tickState"]["ticksDayCounter"].get<uint16_t>();
